Add `styled-components` & Its Global Style
Add `styled-components` and its Global Style so that the developer can
style the component in the component to make it easier to see which
component is being styled. Style Components can be used to describe
better what the component is supposed to be used for, making the
component easier to read.

Ref:

https://www.npmjs.com/package/styled-components